refactor(api): Migrate Sentry integration to monitoring module
- Moved Sentry integration from the config module to a new monitoring module for better organization.
- Updated imports in main application, exception handler, and worker tasks to reflect the new module structure.
- Removed the legacy Sentry configuration file to streamline the codebase.
