+++ components/ecoindex/monitoring/__init__.py
@@ -1,0 +1,15 @@
+from ecoindex.monitoring.sentry import (
+    INTERNAL_SERVER_ERROR_STATUS,
+    capture_internal_error,
+    capture_task_failure,
+    get_sentry_environment,
+    init_sentry,
+)
+
+__all__ = [
+    "INTERNAL_SERVER_ERROR_STATUS",
+    "capture_internal_error",
+    "capture_task_failure",
+    "get_sentry_environment",
+    "init_sentry",
+]
